chore: moved cypress cloud tasks test contents to shared
